% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_SCAN.r
\name{Q_SCAN}
\alias{Q_SCAN}
\title{Quadratic SCAN (Q-SCAN) statistic based procedure}
\usage{
Q_SCAN(
  genotype,
  phenotype,
  X,
  family,
  Lmax,
  Lmin,
  steplength = 1,
  times = 2000,
  alpha = 0.05,
  f = 0
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence, where n is the sample
size and p is the number of variants.}

\item{phenotype}{an n*1 phenotype vector, where n is the sample size.}

\item{X}{an n*q covariates matrix, where n is the sample size and q is the number of covariates.}

\item{family}{a description of the error distribution and link function to be used
in the model. This can be a character string naming a family function, a family
function or the result of a call to a family function. (See \code{\link{family}} for details of family functions).
Can be either "gaussian" for continuous phenotype or "binomial" for binary phenotype.}

\item{Lmax}{maximum number of variants in searching windows.}

\item{Lmin}{minimum number of variants in searching windows.}

\item{steplength}{difference of number of variants in searching windows, that is, the number of variants in
searching windows are Lmin, Lmin+steplength, Lmin+steplength,...,Lmax. (Default is 1).}

\item{times}{a number of pesudo-residuals (default = 2000).}

\item{alpha}{familty-wise/genome-wide significance level. (Default is 0.05).}

\item{f}{an overlap fraction, which controls for the overlapping proportion of of detected regions. For example,
when f=0, the detected regions are non-overlapped with each other,
and when f=1, we keep every susceptive region as detected regions. (Default is 0.)}
}
\value{
The function returns a list with the following members:

\code{SCAN_res}:   A matrix that summarized the significant region detected by Q-SCAN.
The first column is the quadratic scan statistic of the detected region.
The next two columns are the location of the detected region (in sense of variants order).
The last column is the family-wise/genome-wide error rate of the detected region.
The result (0,0,0,1) means there is no significant region.

\code{SCAN_top1}:  A vector of length 4 which summarized the top 1 region detected by Q-SCAN.
The first element is the quadratic scan statistic of the detected region
The next two elements are the location of the detected region (in sense of variants order).
The last element is the family-wise/genome-wide p-value.

\code{SCAN_thres}: Empirical threshold of Q-SCAN for controlling the family-wise type I error at alpha level.

\code{SCAN_thres_boot}: A vector of Monte Carlo simulation sample for generating the empirical threshold. The 1-alpha quantile of this vector is
the empirical threshold.
}
\description{
The \code{Q_SCAN} function takes in genotype, phenotype and covariates and detect the association between a
quantitative/dichotomous phenotype and a variant-set in a sequence by using quadratic scan statistic based procedure.
}
